Corrigido inicialização do módulo de currencies
-- O módulo de currencies estava sendo inicializado depois do carregamente do sapper, que estava causando as stores serem inicializadas com valores inválidos. Para corrigir isso o carregamento do sapper foi atrasado no cliente e no servidor
-- Tanto o servidor como o cliente só serão inicializados depois de o módulo de currencies ser coretamente populado
-- Removido carregamento do módulo de currencies da função de preload do _layout principal
-- O cache de currencies do servidor agora é o módulo de currencies, então a rota /currencies retorna o objeto armazenado lá para responder o request
-- Removido woraround da página de balances para não crashar a pǵina por a store não ter o formato adequado
