+++ client/package.json
@@ -9,7 +9,7 @@
     "build:clean": "del-cli ./__sapper__/build",
     "build:sapper": "sapper build",
     "build:repository": "copyfiles -E static/** static/**/* package* Dockerfile __sapper__/build/",
-    "build:fix-configs": "rexreplace \"__sapper__/build\" \".\" __sapper__/build/server/server.js && rexreplace \"cd __sapper__/build && \" \"\" __sapper__/build/package.json",
+    "build:fix-configs": "rexreplace \"__sapper__/build\" \".\" __sapper__/build/server/*.js && rexreplace \"cd __sapper__/build && \" \"\" __sapper__/build/package.json",
     "export": "sapper export",
     "cy:run": "cypress run",
     "cy:open": "cypress open",
