+++ client/src/client.js
@@ -1,6 +1,18 @@
+import axios from 'axios'
 import * as sapper from '@sapper/app'
+import { init } from './utils/currencies'
 import { connect } from './utils/websocket'
 
-sapper.start({
-	target: document.querySelector('#sapper')
-}).then(connect)
+// Conecta com o websocket
+connect()
+
+// Inicializa o módulo de currencies
+axios.get('/currencies').then(res => {
+	init(res.data)
+}).then(() => {
+	return sapper.start({
+		target: document.querySelector('#sapper')
+	})
+}).catch(err => {
+	console.error('Error initializing client:', err)
+})
